[BACKEND] Develop complete CRUD endpoints for /api/transacoes/consumo.
Fixes AB#365
