+++ src/main/java/com/essjr/DinMonex/auth/dtos/LoginRequestDTO.java
@@ -1,0 +1,17 @@
+package com.essjr.DinMonex.auth.dtos;
+
+public class LoginRequestDTO {
+    private String email;
+    private String password;
+
+    public LoginRequestDTO(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+
+    // Getters e Setters
+    public String getEmail() { return email; }
+    public void setEmail(String email) { this.email = email; }
+    public String getPassword() { return password; }
+    public void setPassword(String password) { this.password = password; }
+}
