+++ src/main/java/com/essjr/DinMonex/auth/dtos/LoginResponseDTO.java
@@ -1,0 +1,13 @@
+package com.essjr.DinMonex.auth.dtos;
+
+public class LoginResponseDTO {
+    private String token;
+
+    public LoginResponseDTO(String token) {
+        this.token = token;
+    }
+
+    // Getter e Setter
+    public String getToken() { return token; }
+    public void setToken(String token) { this.token = token; }
+}
