--- src/main/java/com/essjr/DinMonex/dtos/ApiResponseDTO.java
@@ -1,16 +1,0 @@
-package com.essjr.DinMonex.dtos;
-
-public class ApiResponseDTO {
-    private String message;
-
-    public ApiResponseDTO(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
