--- src/main/java/com/essjr/DinMonex/dtos/AppUserDTO.java
@@ -1,12 +1,0 @@
-package com.essjr.DinMonex.dtos;
-
-import com.essjr.DinMonex.model.enuns.AppUserRole;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-public record AppUserDTO(@NotNull Long id,
-                         @NotBlank String name,
-                         @NotBlank @Email String email,
-                         AppUserRole role) {
-}
