--- src/main/java/com/essjr/DinMonex/model/enuns/AppUserRole.java
@@ -1,6 +1,0 @@
-package com.essjr.DinMonex.model.enuns;
-
-public enum AppUserRole {
-    REGULAR,
-    ADMIN
-}
