--- src/main/java/com/essjr/DinMonex/model/enuns/TransactionType.java
@@ -1,5 +1,0 @@
-package com.essjr.DinMonex.model.enuns;
-
-public enum TransactionType {
-    CONSUMPTION, CREDIT_CARD, DEBT
-}
