--- src/main/java/com/essjr/DinMonex/repositories/AppUserRepository.java
@@ -1,11 +1,0 @@
-package com.essjr.DinMonex.repositories;
-
-import com.essjr.DinMonex.model.AppUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface AppUserRepository extends JpaRepository<AppUser, Long> {
-
-    Optional<AppUser> findByEmail(String email);
-}
