--- src/main/java/com/essjr/DinMonex/services/JpaUserDetailsService.java
@@ -1,24 +1,0 @@
-package com.essjr.DinMonex.services;
-
-import com.essjr.DinMonex.repositories.AppUserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class JpaUserDetailsService implements UserDetailsService {
-    private final AppUserRepository appUserRepository;
-
-    public JpaUserDetailsService(AppUserRepository appUserRepository) {
-        this.appUserRepository = appUserRepository;
-    }
-
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return appUserRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found with email: " + username));
-    }
-
-}
