+++ src/main/java/com/essjr/DinMonex/shared/ApiResponseDTO.java
@@ -1,0 +1,16 @@
+package com.essjr.DinMonex.shared;
+
+public class ApiResponseDTO {
+    private String message;
+
+    public ApiResponseDTO(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
