+++ src/main/java/com/essjr/DinMonex/transaction/enuns/TransactionType.java
@@ -1,0 +1,5 @@
+package com.essjr.DinMonex.transaction.enuns;
+
+public enum TransactionType {
+    CONSUMPTION, CREDIT_CARD, DEBT
+}
