+++ src/main/java/com/essjr/DinMonex/user/AppUserRepository.java
@@ -1,0 +1,10 @@
+package com.essjr.DinMonex.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface AppUserRepository extends JpaRepository<AppUser, Long> {
+
+    Optional<AppUser> findByEmail(String email);
+}
