+++ src/main/java/com/essjr/DinMonex/user/enums/AppUserRole.java
@@ -1,0 +1,6 @@
+package com.essjr.DinMonex.user.enums;
+
+public enum AppUserRole {
+    REGULAR,
+    ADMIN
+}
