+++ src/test/java/com/essjr/DinMonex/repositories/AppUserRepositoryTest.java
@@ -1,14 +1,13 @@
 package com.essjr.DinMonex.repositories;
 
-import com.essjr.DinMonex.model.AppUser;
-import com.essjr.DinMonex.model.enuns.AppUserRole;
+import com.essjr.DinMonex.user.AppUser;
+import com.essjr.DinMonex.user.AppUserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.containers.PostgreSQLContainer;
@@ -18,7 +17,6 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 
 /**
@@ -70,7 +68,7 @@ class AppUserRepositoryTest {
         appUser.setName("Utilizador Teste");
         appUser.setEmail("teste@email.com");
         appUser.setPassword("hash123");
-        appUser.setRole(AppUserRole.REGULAR);
+        appUser.setRole(AppUser.AppUserRole.REGULAR);
     }
 
     @Test
