+++ src/test/java/com/essjr/DinMonex/services/JpaUserDetailsServiceTest.java
@@ -1,8 +1,8 @@
 package com.essjr.DinMonex.services;
 
-import com.essjr.DinMonex.model.AppUser;
-import com.essjr.DinMonex.model.enuns.AppUserRole;
-import com.essjr.DinMonex.repositories.AppUserRepository;
+import com.essjr.DinMonex.auth.JpaUserDetailsService;
+import com.essjr.DinMonex.user.AppUser;
+import com.essjr.DinMonex.user.AppUserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,7 +17,6 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -35,7 +34,7 @@ class JpaUserDetailsServiceTest { // <-- MUDANÇA DE NOME AQUI
 
     @BeforeEach
     void setUp() {
-        testUser = new AppUser(1L, "Test User", userEmail, "password123", AppUserRole.REGULAR);
+        testUser = new AppUser(1L, "Test User", userEmail, "password123", AppUser.AppUserRole.REGULAR);
     }
 
     @Test
